[action] [PR:23478] [devutil]: Add pmon readiness check and refactor DPU NAT utilities (#23954)
### Description of PR

Summary:
After an image upgrade and reboot on SmartSwitch, DPUs may not be
immediately available when `sonic-dpu-mgmt-traffic.sh` runs. This causes
NAT enablement failures in `enable_nat_for_dpuhosts()` because the pmon
container (required to manage DPUs) is not yet running.

This PR:
1. Refactors DPU NAT utilities out of `sonic.py` into a new dedicated
`dpu_utils.py` module.
2. Adds a `is_pmon_running()` readiness check that polls for the pmon
container (up to 300s, every 30s) before attempting NAT configuration.

### Type of change

- [x] Bug fix
- [x] Testbed and Framework(new/improvement)
- [ ] New Test case
 - [ ] Skipped for non-supported platforms
- [ ] Test case improvement

### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505
- [x] 202511

### Approach

#### What is the motivation for this PR?
During SmartSwitch image upgrades, `enable_nat_for_dpuhosts()` runs
immediately after reboot completes. However, the pmon container—required
for DPU management—needs additional time to initialize. Without pmon
running, `sonic-dpu-mgmt-traffic.sh` fails because DPUs are not yet
detected on the platform.

#### How did you do it?
1. **Created `ansible/devutil/devices/dpu_utils.py`** — Moved
`is_nat_enabled_for_dpu()`, `get_dpu_name_ssh_port_dict()`, and
`enable_nat_for_dpuhosts()` out of `sonic.py` into this new module to
improve code organization and reusability.
2. **Added `is_pmon_running()`** — A new function that polls `docker ps
| grep pmon` up to 10 times (every 30s, max 300s) waiting for the pmon
container to come up. If pmon is not running after the timeout, NAT
configuration is skipped for that host.
3. **Integrated pmon check into `enable_nat_for_dpuhosts()`** — Before
configuring NAT on each NPU host, the function now waits for pmon
readiness.
4. **Updated imports** in `.azure-pipelines/testbed_health_check.py` and
`.azure-pipelines/upgrade_image.py` to use the new `dpu_utils` module.

#### How did you verify/test it?
- Analyzed failure logs from SmartSwitch testbed showing the timing
issue where pmon was not ready
- Verified the NAT enablement succeeds after pmon container is up
- Code review of pmon readiness check implementation

#### Any platform specific information?
This fix applies to SmartSwitch platforms with DPUs (e.g., MSN4280).

#### Supported testbed topology if it is a new test case?
N/A — This is a bug fix and refactor, not a new test case.

### Documentation
N/A — Internal infrastructure fix, no user-facing documentation
required.
